requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.12.5
chardet==4.0.0
click
dnspython
email-validator
fastapi
gevent
graphene
graphql-core
graphql-relay
greenlet
grequests
h11
httptools
idna
itsdangerous
Jinja2
MarkupSafe
orjson==3.5.2
promise==2.3
pydantic==1.8.2
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
requests-cache==0.6.3
Rx==1.6.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
url-normalize==1.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
zope.event==4.5.0
zope.interface==5.4.0
